Add Post service layer

+++ Posterr/Services/UserService/UserService.swift
@@ -8,5 +8,5 @@
 import Foundation
 
 protocol UserService {
-  func userById(_ id: Int) -> User?
+  func userById(_ id: Int64) -> User?
 }
